+++ deprecated/ui-common/src/services/bridge/types.ts
@@ -16,6 +16,7 @@ export enum BridgingNetwork {
   Osmosis = 'osmosis',
   Persistence = 'Persistence',
   Terra = 'terra',
+  Moonbeam = 'moonbeam',
 }
 
 export const MintScanExplorerUrl = {