+++ packages/sdk-ui-ts/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.31](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@0.0.30...@injectivelabs/sdk-ui-ts@0.0.31) (2022-06-03)
+
+
+### Bug Fixes
+
+* exports ([bc77a41](https://github.com/InjectiveLabs/injective-ts/commit/bc77a4179fa642f3fd2fc38f0ac7c02eb2965f1e))
+
+
+
+
+
+## [0.0.30](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@0.0.29...@injectivelabs/sdk-ui-ts@0.0.30) (2022-06-03)
+
+
+### Bug Fixes
+
+* update bridge networks ([b7dd35b](https://github.com/InjectiveLabs/injective-ts/commit/b7dd35b7addde30123c73894f171cbcfe2d44450))
+
+
+
+
+
 ## [0.0.29](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@0.0.28...@injectivelabs/sdk-ui-ts@0.0.29) (2022-06-03)