+++ packages/wallet-ts/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.12](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.11...@injectivelabs/wallet-ts@0.0.12) (2022-06-03)
+
+
+### Bug Fixes
+
+* axelar api moonbeam ([aee6731](https://github.com/InjectiveLabs/injective-ts/commit/aee673125984bf74bb2e8b3483935eb1a46b8f6d))
+
+
+
+
+
+## [0.0.11](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.10...@injectivelabs/wallet-ts@0.0.11) (2022-06-03)
+
+
+### Bug Fixes
+
+* exports ([bc77a41](https://github.com/InjectiveLabs/injective-ts/commit/bc77a4179fa642f3fd2fc38f0ac7c02eb2965f1e))
+
+
+
+
+
+## [0.0.10](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.9...@injectivelabs/wallet-ts@0.0.10) (2022-06-03)
+
+
+### Bug Fixes
+
+* package export ([cef1d5a](https://github.com/InjectiveLabs/injective-ts/commit/cef1d5ab504de5a7d085860dda256a793efd82aa))
+
+
+
+
+
+## [0.0.9](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.8...@injectivelabs/wallet-ts@0.0.9) (2022-06-03)
+
+
+### Bug Fixes
+
+* wallet ts export ([3d64721](https://github.com/InjectiveLabs/injective-ts/commit/3d64721e9538e40d19f932b286337132f2d0ce14))
+
+
+
+
+
+## [0.0.8](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.7...@injectivelabs/wallet-ts@0.0.8) (2022-06-03)
+
+
+### Bug Fixes
+
+* export ([6522c42](https://github.com/InjectiveLabs/injective-ts/commit/6522c42e1301004166281dc04d6e0be083cf9cda))
+
+
+
+
+
 ## [0.0.7](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.6...@injectivelabs/wallet-ts@0.0.7) (2022-06-03)